requirements.txt
streamlit
python-docx
python-dotenv
langchain
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
